Basic maltego transforms complete. 1. Port Scan 2. PMA Scan 3. Admin Scan 4. CMS Detection Scan
