--- lib/tools/nettacker_transforms/src/nettacker_transforms/resources/etc/nettacker_transforms.conf
@@ -1,43 +1,0 @@
-# This is a sample configuration file that you can use to configure global options for your Canari project. Common use-
-# cases for this configuration file include API keys, performance options, etc. Below are some sample sections with
-# example option types. You can remove these sections completely with no effect.
-#
-# Canari's configuration parser automatically marshals option values into their appropriate types when setting or
-# getting options. The following option value types are supported:
-#
-# List option:
-# foo = ["this", "is", "a", "list"]
-#
-# String option:
-# foo = this is a string
-# foo = this\, is\, a\, comma-separated string
-#
-# Multiline string option:
-# foo: this is a very long
-#    string
-#
-# Interpolated string option with environment variables:
-# foo = ${PATH}:/usr/local/bin
-# foo = ${HOME}/mydata.txt
-#
-# Interpolated string option with other option variables:
-# foo = %(bar)/foo
-# bar = baz
-#
-# Integer option:
-# foo = 1
-#
-# Float option:
-# foo = 2.2
-#
-# Python object option (only for local transforms):
-# foo = object://os.path/join
-
-
-[nettacker_transforms.local]
-
-# TODO: put local transform options here
-
-[nettacker_transforms.remote]
-
-# TODO: put remote transform options here
