+++ lib/transactions/maltego/nettacker_transforms/src/canari.conf
