+++ lib/transactions/maltego/nettacker_transforms/src/nettacker_transforms/resources/__init__.py
