+++ lib/transactions/maltego/nettacker_transforms/src/nettacker_transforms/resources/etc/nettacker_transforms.conf
