+++ lib/transactions/maltego/nettacker_transforms/src/nettacker_transforms/resources/maltego/entities.mtz
