+++ lib/transactions/maltego/nettacker_transforms/src/nettacker_transforms/transforms/common/entities.py
